[WIP] Update all dependencies

+++ src/gui/element/mod.rs
@@ -10,7 +10,7 @@ pub static SMALLER_FONT_SIZE: u16 = 12;
 pub static DEFAULT_FONT_SIZE: u16 = 14;
 pub static DEFAULT_HEADER_FONT_SIZE: u16 = 24;
 pub static DEFAULT_SUB_HEADER_FONT_SIZE: u16 = 18;
-pub static DEFAULT_PADDING: u16 = 10;
+pub static DEFAULT_PADDING: f32 = 10.0;
 
-pub static BUTTON_WIDTH: u16 = 84;
-pub static BUTTON_HEIGHT: u16 = 16;
+pub static BUTTON_WIDTH: f32 = 84.0;
+pub static BUTTON_HEIGHT: f32 = 16.0;
